// MovieApi/src/app/models/movie-item.interface.ts
// Generated by https://quicktype.io

export interface MovieResponse {
    page:          number;
    results:       Movie[];
    total_pages:   number;
    total_results: number;
}

export interface Movie {
    adult:             boolean;
    backdrop_path:     string;
    genre_ids:         number[];
    id:                number;
    original_language: OriginalLanguage;
    original_title:    string;
    overview:          string;
    popularity:        number;
    poster_path:       string;
    release_date:      string;
    title:             string;
    video:             boolean;
    vote_average:      number;
    vote_count:        number;
}

export enum OriginalLanguage {
    En = "en",
    Hi = "hi",
}